+++ router.go
@@ -412,10 +412,6 @@ func (r *Router) requestHandler(match *routeMatch, w http.ResponseWriter, rawReq
 
 // finalize the request's life-cycle.
 func (r *Router) finalize(response *Response, request *Request) {
-	if response.Hijacked() { // TODO test hijacked
-		return
-	}
-
 	if response.empty {
 		if response.status == 0 {
 			// If the response is empty, return status 204 to
@@ -424,12 +420,11 @@ func (r *Router) finalize(response *Response, request *Request) {
 		} else if statusHandler, ok := r.statusHandlers[response.status]; ok {
 			// Status has been set but body is empty.
 			// Execute status handler if exists.
-			// TODO status handlers on hijacked connections?
 			statusHandler(response, request)
 		}
 	}
 
-	if !response.wroteHeader {
+	if !response.wroteHeader && !response.hijacked {
 		response.WriteHeader(response.status)
 	}